+++ src/http_utils.py
@@ -1,0 +1,112 @@
+import base64
+import hmac as _hmac
+import json
+import re
+import traceback
+from urllib.parse import urlparse
+
+from workers import Response
+
+
+def capture_exception(exc: Exception, req=None, _env=None, where: str = ""):
+    """Best-effort exception logging with full traceback and request context."""
+    try:
+        payload = {
+            "level": "error",
+            "where": where or "unknown",
+            "error_type": type(exc).__name__,
+            "error": str(exc),
+            "traceback": "".join(traceback.format_exception(type(exc), exc, exc.__traceback__)),
+        }
+        if req:
+            payload["request"] = {
+                "method": req.method,
+                "url": req.url,
+                "path": urlparse(req.url).path,
+            }
+        print(json.dumps(payload))
+    except Exception:
+        pass
+
+
+CORS = {
+    "Access-Control-Allow-Origin": "*",
+    "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, OPTIONS",
+    "Access-Control-Allow-Headers": "Content-Type, Authorization",
+}
+
+
+def json_resp(data, status: int = 200):
+    return Response(
+        json.dumps(data),
+        status=status,
+        headers={"Content-Type": "application/json", **CORS},
+    )
+
+
+def ok(data=None, msg: str = "OK"):
+    body = {"success": True, "message": msg}
+    if data is not None:
+        body["data"] = data
+    return json_resp(body, 200)
+
+
+def err(msg: str, status: int = 400):
+    return json_resp({"error": msg}, status)
+
+
+async def parse_json_object(req):
+    """Parse request JSON and ensure payload is an object/dict."""
+    try:
+        text = await req.text()
+        body = json.loads(text)
+    except Exception:
+        return None, err("Invalid JSON body")
+
+    if not isinstance(body, dict):
+        return None, err("JSON body must be an object", 400)
+
+    return body, None
+
+
+def clean_path(value: str, default: str = "/admin") -> str:
+    """Normalize an env-provided path into a safe absolute URL path."""
+    raw = (value or "").strip()
+    if not raw:
+        return default
+    parsed = urlparse(raw)
+    path = (parsed.path or raw).strip()
+    if not path.startswith("/"):
+        path = "/" + path
+    path = re.sub(r"/+", "/", path)
+    if len(path) > 1 and path.endswith("/"):
+        path = path[:-1]
+    return path or default
+
+
+def unauthorized_basic(realm: str = "Alpha One Labs Admin"):
+    return Response(
+        "Authentication required",
+        status=401,
+        headers={"WWW-Authenticate": f'Basic realm="{realm}"', **CORS},
+    )
+
+
+def is_basic_auth_valid(req, env) -> bool:
+    username = (getattr(env, "ADMIN_BASIC_USER", "") or "").strip()
+    password = (getattr(env, "ADMIN_BASIC_PASS", "") or "").strip()
+    if not username or not password:
+        return False
+
+    auth = req.headers.get("Authorization") or ""
+    if not auth.lower().startswith("basic "):
+        return False
+
+    try:
+        raw = auth.split(" ", 1)[1].strip()
+        decoded = base64.b64decode(raw).decode("utf-8")
+        user, pwd = decoded.split(":", 1)
+    except Exception:
+        return False
+
+    return _hmac.compare_digest(user, username) and _hmac.compare_digest(pwd, password)
